Polish two-way tables

eda_pol Polishes two-way tables using median, means, or any customizable functions.

eda_pol(
  x,
  row = NULL,
  col = NULL,
  val = NULL,
  stat = median,
  plot = TRUE,
  eps = 0.01,
  maxiter = 5,
  sort = FALSE,
  p = 1,
  tukey = FALSE,
  col.quant = FALSE,
  colpal = "RdYlBu",
  adj.mar = FALSE,
  res.size = 1,
  row.size = 1,
  col.size = 1,
  res.txt = TRUE,
  label.txt = TRUE
)

Arguments

x

A three column data frame

row

Name of column assigned to the row effect

col

Name of column assigned to the column effect

val

Name of column assigned to the response variable

stat

Polishing statistic (default is median)

plot

Boolean determining if an output plot should be generated

eps

Convergence tolerance parameter

maxiter

Maximum number of iterations

sort

Boolean determining if the effects row/columns should be sorted

p

Re-expression power parameter

tukey

Boolean determining if Tukey's power transformation should used. If FALSE, the Box-Cox transformation is adopted.

col.quant

Boolean determining if a quantile classification scheme should be used

colpal

Color palette to adopt

adj.mar

Boolean determining if margin width needs to accomodate labels

res.size

Size of residual values in plot [0-1]

row.size

Size of row effect values in plot [0-1]

col.size

Size of column effect values in plot [0-1]

res.txt

Boolean determining if values should be added to plot

label.txt

Boolean determining if margin and column labels should be plotted

Details

The function performs a polish on a two way table. By default, it applies a median polish, but other statistic such as the mean can be passed to the function via the stat = parameter. The function returns a list of row/column effects along with global and residual values. It will also generate a colored table if plot = TRUE. Returns a list of class eda_polish

Examples

df <- data.frame(region =  rep( c("NE", "NC", "S", "W"), each = 5),
edu = rep( c("ed8", "ed9to11", "ed12", "ed13to15", "ed16"), 4),
perc = c(25.3, 25.3, 18.2, 18.3, 16.3, 32.1, 29, 18.8,
        24.3, 19, 38.8, 31, 19.3, 15.7, 16.8, 25.4, 21.1, 20.3, 24, 17.5))

eda_pol(df, row = "region", col = "edu", val = "perc", plot = FALSE)
#> $long
#>         edu region     perc
#> 1      ed12     NC -3.15625
#> 2      ed12     NE  0.00625
#> 3      ed12      W  0.29375
#> 4      ed12      S -0.00625
#> 5  ed13to15     NE -1.11875
#> 6  ed13to15      S -4.83125
#> 7  ed13to15     NC  1.11875
#> 8  ed13to15      W  2.76875
#> 9      ed16      S  0.00000
#> 10     ed16     NC -0.45000
#> 11     ed16      W  0.00000
#> 12     ed16     NE  0.61250
#> 13      ed8     NC  1.51875
#> 14      ed8     NE -1.51875
#> 15      ed8      S 10.86875
#> 16      ed8      W -3.23125
#> 17  ed9to11     NE  0.03125
#> 18  ed9to11     NC -0.03125
#> 19  ed9to11      S  4.61875
#> 20  ed9to11      W -5.98125
#> 
#> $wide
#>         eff     ed12 ed13to15    ed16      ed8  ed9to11
#> eff 20.8500 -1.19375  0.03125 -3.7000  7.43125  5.88125
#> NC   2.3000 -3.15625  1.11875 -0.4500  1.51875 -0.03125
#> NE  -1.4625  0.00625 -1.11875  0.6125 -1.51875  0.03125
#> S   -0.3500 -0.00625 -4.83125  0.0000 10.86875  4.61875
#> W    0.3500  0.29375  2.76875  0.0000 -3.23125 -5.98125
#> 
#> $row
#>   region  effect
#> 1     NC  2.3000
#> 2     NE -1.4625
#> 3      S -0.3500
#> 4      W  0.3500
#> 
#> $col
#>        edu   effect
#> 1     ed12 -1.19375
#> 2 ed13to15  0.03125
#> 3     ed16 -3.70000
#> 4      ed8  7.43125
#> 5  ed9to11  5.88125
#> 
#> $global
#> [1] 20.85
#> 
#> $iter
#> [1] 5
#> 
#> $cv
#>        perc     row      col            cv
#> 1  -3.15625  2.3000 -1.19375 -0.1316846523
#> 2  -0.00625 -0.3500 -1.19375  0.0200389688
#> 3   0.00625 -1.4625 -1.19375  0.0837342626
#> 4   0.29375  0.3500 -1.19375 -0.0200389688
#> 5  -4.83125 -0.3500  0.03125 -0.0005245803
#> 6   1.11875  2.3000  0.03125  0.0034472422
#> 7   2.76875  0.3500  0.03125  0.0005245803
#> 8  -1.11875 -1.4625  0.03125 -0.0021919964
#> 9  -0.45000  2.3000 -3.70000 -0.4081534772
#> 10  0.61250 -1.4625 -3.70000  0.2595323741
#> 11  0.00000  0.3500 -3.70000 -0.0621103118
#> 12  0.00000 -0.3500 -3.70000  0.0621103118
#> 13 -1.51875 -1.4625  7.43125 -0.5212567446
#> 14 10.86875 -0.3500  7.43125 -0.1247452038
#> 15  1.51875  2.3000  7.43125  0.8197541966
#> 16 -3.23125  0.3500  7.43125  0.1247452038
#> 17 -0.03125  2.3000  5.88125  0.6487709832
#> 18  0.03125 -1.4625  5.88125 -0.4125337230
#> 19  4.61875 -0.3500  5.88125 -0.0987260192
#> 20 -5.98125  0.3500  5.88125  0.0987260192
#> 
#> $power
#> [1] 1
#> 
#> $IQ_row
#> [1] 0.4177057
#> 
#> $IQ_col
#> [1] 2.016388
#> 
#> attr(,"class")
#> [1] "eda_polish"