Generate Data with Specific Skewness and Kurtosis

This function generates a dataset with a specified skewness and kurtosis using Fleishman's polynomial transformation.

Usage

eda_sim(n, skewness = 0, kurtosis = 0, ...)

Arguments

n

Integer. Number of data points to generate.

skewness

Numeric. Desired skewness of the generated data. Default is 0 (symmetric distribution).

kurtosis

Numeric. Desired kurtosis of the generated data. Default is 0 (normal distribution).

...

Not used.

Value

A numeric vector of length n containing the generated data.

Details

The function uses Fleishman's polynomial transformation of the form: $$Y = a + bX + cX^2 + dX^3$$ where a, b, c, and d are coefficients determined to approximate the specified skewness and kurtosis. The coefficients are solved using a numerical optimization approach based on minimizing the residuals of Fleishman's equations.

References

Fleishman, A. I. (1978). A method for simulating non-normal distributions. Psychometrika, 43, 521–532.

Examples

 # A normal distribution
 x <- eda_sim(5000, 0, 0)
#> Error in eda_sim(5000, 0, 0): could not find function "eda_sim"
 hist(x)
#> Error: object 'x' not found

 # A right-skewed distribution
 x <- eda_sim(5000, 2, 0)
#> Error in eda_sim(5000, 2, 0): could not find function "eda_sim"
 hist(x)
#> Error: object 'x' not found

 # A left-skewed distribution
 x <- eda_sim(5000, -2, 0)
#> Error in eda_sim(5000, -2, 0): could not find function "eda_sim"
 hist(x)
#> Error: object 'x' not found

 # A uniform distribution
 # Note that this is unbounded which may result in outliers
 set.seed(21)
 x <- eda_sim(5000, 0, -10)
#> Error in eda_sim(5000, 0, -10): could not find function "eda_sim"
 hist(x, breaks = 20)
#> Error: object 'x' not found

 # A "peaky" distribution
  set.seed(12)
  x <- eda_sim(5000, 0, 10)
#> Error in eda_sim(5000, 0, 10): could not find function "eda_sim"
  hist(x, breaks = 20)
#> Error: object 'x' not found

  # Distributions can be modified so as to share the same mean and
  # variances. For example, to have x1 (a right skewed distribution) share
  # the  same mean and variance as x2 (a left skewed distribution), type:
  set.seed(432)
  x1 <- eda_sim(10000, 0.5, 0)
#> Error in eda_sim(10000, 0.5, 0): could not find function "eda_sim"
  x2 <- eda_sim(10000, -0.5, 0)
#> Error in eda_sim(10000, -0.5, 0): could not find function "eda_sim"

  x1_u <- mean(x1)
#> Error: object 'x1' not found
  x1_sd <- sd(x1)
#> Error: object 'x1' not found
  x2_sd <- sqrt(var(x2))
#> Error: object 'x2' not found
  x1b <- (x1 - x1_u) / x1_sd * x2_sd + mean(x2)
#> Error: object 'x1' not found
  mean(x1b)
#> Error: object 'x1b' not found
  mean(x2)
#> Error: object 'x2' not found
  var(x1b)
#> Error: object 'x1b' not found
  var(x2)
#> Error: object 'x2' not found

eda_dens(x1b, x2)
#> Error: object 'x1b' not found