yarn dataset

Table 4-19 of Exploring data tables, trends, and shapes shows data from a textile experiment where specimens of worsted yarn were repeatedly loaded until they broke. The table shows the resulting number of cycles to failure for the yarn under different conditions related to the length of the specimen, the amplitude of the loading, and the load applied.

Usage

data(yarn)

Format

A data frame with the following variables:

Length

Length effect

Amplitude

Amplitude effect

Load

Load effect

Cycles

Response variable

References

Hoaglin, David C. and Mosteller, Frederick and Tukey, John W. (1985). Exploring data tables, trends, and shapes. Wiley.